requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
fastapi==0.115.6
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
jinja2==3.1.4
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
pydantic==2.10.3
    # via
    #   fastapi
    #   sqlmodel
pydantic-core==2.27.1
    # via pydantic
pyjwt==2.10.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   uvicorn
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   sqlmodel
sqlmodel==0.0.22
    # via -r requirements.in
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.32.1
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.0
    # via uvicorn
websockets==14.1
    # via uvicorn
